remove canvas from ecosystem model as not used
