+++ src/domain/context/ecosystem-model/dto/ecosystem-model.dto.create.ts
