+++ src/domain/context/ecosystem-model/ecosystem-model.entity.ts
@@ -1,8 +1,7 @@
-import { Column, Entity, JoinColumn, OneToMany, OneToOne } from 'typeorm';
+import { Column, Entity, OneToMany } from 'typeorm';
 import { IEcosystemModel } from '@domain/context/ecosystem-model/ecosystem-model.interface';
 import { ActorGroup } from '@domain/context/actor-group/actor-group.entity';
 import { AuthorizableEntity } from '@domain/common/entity/authorizable-entity';
-import { Canvas } from '@domain/common/canvas';
 @Entity()
 export class EcosystemModel
   extends AuthorizableEntity
@@ -16,14 +15,6 @@ export class EcosystemModel
     cascade: true,
   })
   actorGroups?: ActorGroup[];
-
-  @OneToOne(() => Canvas, {
-    eager: true,
-    cascade: true,
-    onDelete: 'SET NULL',
-  })
-  @JoinColumn()
-  canvas?: Canvas;
 
   // The restricted actor group names at the Opportunity level
   restrictedActorGroupNames: string[];
