+++ src/domain/context/ecosystem-model/index.ts
@@ -1,4 +1,4 @@
 export * from './ecosystem-model.entity';
 export * from './ecosystem-model.interface';
-export * from './ecosystem-model.dto.create';
-export * from './ecosystem-model.dto.update';
+export * from './dto/ecosystem-model.dto.create';
+export * from './dto/ecosystem-model.dto.update';
